only find files once
This is a bit of a weird one, but hear me out.

The `scanner.scan()` files has multiple responsibilities:

1. Figure out files and globs from the file system based on auto content
   detection.
2. Figure out files and globs from the files system based on `@source`
   (GlobEntry) directives.
3. Actually scan those files to extract candidates from them.

This means that if you want to access `scanner.files` or `scanner.globs`
(to setup watchers, or register them with PostCSS) that you _have_ to
call `scanner.scan()` first.

Setting up a watcher (e.g.: inside of the CLI) is an async action. So it
_could_ be that a file changed in between the `scanner.scan()` and
`await setupWatchers()` which could be a race condition.

Another issue is that we call `let candidates = scanner.scan()` but we
only need it later, but moving it closer to where we need it breaks the
`scanner.files` and `scanner.globs` code...

So instead, we want to split up the responsibility of `scan()` to first
figure out all files/globs then scan the files for candidates.

If we do that, we have to introduce a new function like
`scanner.prepareFilesAndGlobs()` which feels silly because you always
have to do that before accessing `scanner.files` or `scanner.globs`. In
fact, you also have to do that before calling `scanner.scan()`.

Another solution is to make `scanner.files()` and `scanner.globs()`
functions. Whenever these are called we do the discovery of those files.
However, this now means that we do it `n` times instead of once.

To solve this, we can track some state internally to know whether we did
that work already or not.

This brings us to the next issue. If you, for some reason, don't even
need `scanner.files` or `scanner.globs` (e.g.: an initial build without
any watchers) then `scanner.scan()` won't be able to scan the files
because they are not discovered yet.

So instead, we did this:
1. Setup a scanner
2. When accessing `scanner.files`, `scanner.globs`, calling
   `scanner.scan()` or calling `scanner.scanFiles()` we will discover
   the files/globs first.

Internally we do track whether this work has been done before or not.
But it allows us to delay the discovery until the first time you
actually need it. It allows us to not call a separate function (that you
can forget to call).

It does make the getter on the Rust side side-effect-y but that should
still be safe since JavaScript is single threaded.