improve the Tailwind CSS CLI output when nothing changed
When we observe that no new candidates were found, we didn't write to
disk because nothing really changed. There is also no need to
re-optimize (minify / lightningcss) when nothing changed.

But this had a side effect that when nothing changes that you didn't see
any output.

Typically you are changing things, so it's not really a problem. But the
moment you use a class that already existed you also don't get any
output because we have a shared cache.

This improves that by always having some output.