Improve the CLI output when nothing changed (#14351)
When we observe that no new candidates were found, then we can return
early because nothing really changed. There is also no need to
re-optimize (use Lightning CSS) in this case.

But this had a side effect that when no new candidates were detected,
that you didn't see any output either. This feels like nothing is
working from a DX perspective.

Typically you are changing things, so it's not really a problem. But the
moment you use a class that already existed (e.g.: in another file) you
also don't get any output because we have a shared cache.

This PR solves that by always showing the output. But it still doesn't
write to disk if nothing changed.

---------

Co-authored-by: Adam Wathan <adam.wathan@gmail.com>