copy utils from `@tailwindcss/cli`
Copied some utils from `@tailwindcss/cli`. Initially referenced them,
but we also rely on some dependencies that we have to install.

One solution is to move it to an internal package and import it so that
tsup can inline everything. This also has the same 'we need the correct
dependencies' problem.

Another solution is to create a proper `@tailwindcss/cli-utils` (or
similar) package that we can import as a dependency. However, this one
will require us to actually publish the package, and give it a proper
name.