Skip candidates with invalid `theme()` calls (#14437)
Right now when we encounter a candidates with invalid `theme()` calls we
throw an error which stops the build entirely. This is not ideal
because, especially in the case of `node_modules`, if one file in one
package has an error it will stop the build for an entire project and
tracking this down can be quite difficult.

Now, after this PR, any candidates that use `theme(…)` with non-existent
theme keys (e.g. `rounded-[theme(--radius-does-not-exist)]`) will be
skipped instead of breaking the build.

Before:
```html
<div class="underline rounded-[theme(--radius-does-not-exist)]"></div>
```

```css
/* No CSS was generated because an error was thrown */
/* Error: Invalid theme key: --radius-does-not-exist */
```

After:
```html
<div class="underline rounded-[theme(--radius-does-not-exist)]"></div>
```

```css
.underline {
  text-decoration-line: underline;
}
```

---------

Co-authored-by: Adam Wathan <adam.wathan@gmail.com>